ext trigger only for handy1

+++ docker/run.sh
@@ -2,7 +2,10 @@
 
 set -e
 
-echo 1| sudo tee /sys/module/imx296/parameters/trigger_mode > /dev/null 2>&1 || true
+# handy1向けに、外部トリガを有効化
+if [ -e /sys/module/imx296/parameters/trigger_mode ]; then
+    echo 1 | sudo tee /sys/module/imx296/parameters/trigger_mode > /dev/null 2>&1 || true
+fi
 
 # ユーザーごとの設定を書く用のファイルを設置
 if [ ! -e docker/home/.user_config.bash ]; then
